requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
helpers==0.2.0
httplib2==0.22.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
pypng==0.20220715.0
qrcode==7.4.2
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
